% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_dataset.R
\name{analysis_dataset}
\alias{analysis_dataset}
\title{Derive analysis dataset for weighted analyses}
\usage{
analysis_dataset(
  intervention_wts,
  comparator_data,
  wt_col = WT,
  rs_wt_col = WT_RS
)
}
\arguments{
\item{intervention_wts}{A data frame containing individual patient data
from the intervention study, including a column containing the weights
(derived using estimate_weights).}

\item{comparator_data}{A data frame containing the pseudo-individual patient
data for the comparator intervention.}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is WT.}

\item{rs_wt_col}{The name of the rescaled weights column in the data frame
containing the intervention individual patient data and the MAIC propensity
weights. The default is WT_RS.}
}
\value{
A data frame including the intervention data with propensity weights
  and comparator psuedo data with weights=1.
}
\description{
Creates analysis dataset including weighted intervention data and comparator
pseudo data with propensity weights
}
\seealso{
\code{\link{estimate_weights}}
}
