% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_HR_anchored}
\alias{bootstrap_HR_anchored}
\title{Bootstrapping for MAIC weighted hazard ratios for anchored comparison}
\usage{
bootstrap_HR_anchored(
  intervention_data,
  matching,
  ref_treatment,
  i,
  model,
  min_weight = 1e-04
)
}
\arguments{
\item{intervention_data}{A data frame containing individual patient data
from the intervention study.}

\item{matching}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in intervention_data.}

\item{ref_treatment}{Reference treatment for the intervention_data}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{model}{A model formula in the form 'Surv(Time, Event==1) ~ ARM'.
Variable names need to match the corresponding columns in intervention_data.}

\item{min_weight}{A numeric value that defines the minimum weight allowed. 
This value (default 0.0001) will replace weights estimated at 0 in a sample.}
}
\value{
The logHR as a numeric value.
}
\description{
Bootstrapping for MAIC weighted hazard ratios for anchored comparison
}
\details{
This function is intended to be used in conjunction with the
  \code{\link{boot}} function to return the statistic to be
  bootstrapped. In this case by performing MAIC weighting using
  {\link{estimate_weights}} and returning a weighted hazard ratio (HR) from a
  Cox proportional hazards model. This is used as the 'statistic' argument in
  the boot function.
}
\seealso{
\code{\link{estimate_weights}}, \code{\link{boot}}
}
