% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_RR}
\alias{bootstrap_RR}
\title{Bootstrapping for MAIC propensity weighted relative risk HELP PAGE NEEDS UPDATING}
\usage{
bootstrap_RR(intervention_data, comparator_data, matching, i, model)
}
\arguments{
\item{model}{A model formula in the form 'endpoint ~ treatment_var'.
Variable names need to match the corresponding columns in intervention_data}
}
\value{
A data frame of n_sim bootstrapped OR values.
}
\description{
Bootstrapping of the relative risk to capture the uncertainty in the
estimation of MAIC propensity weights using the following general process:
\enumerate{
   \item Sample patients with replacement from the intervention study
   \item Estimate propensity weights for the current sample of patients
   \item Calculate a weighted relative risk
   \item Repeat the steps above N times. Approx. 1000 simulations are usually
   sufficient
}
Refer to \code{\link{OR_summary}} for alternative approaches to utilising
bootstrapped samples
}
\seealso{
\code{\link{analysis_dataset}}, \code{\link{estimate_weights}}, \code{\link{OR_summary}}
}
