% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{check_weights}
\alias{check_weights}
\title{Checking whether optimization has worked}
\usage{
check_weights(
  analysis_data = NULL,
  matching_vars = NULL,
  target_pop_standard = NULL
)
}
\arguments{
\item{analysis_data}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using estimate_weights).}

\item{target_pop_standard}{aggregate characteristics of the comparator trial with the same naming as the analysis_data}

\item{match_covs}{A character vector giving the names of the covariates that were used to estimate weights}
}
\value{
Summary of patient characteristics before and after matching, including ESS and comparator trial aggregate summary
}
\description{
Convenient function to check whether the re-weighted baseline characteristics for
the intervention-treated patients match those aggregate characteristics from the
comparator trial and outputs a summary that can be used for reporting
}
