% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate MAIC propensity weights}
\usage{
estimate_weights(intervention_data, matching_vars, method = "BFGS", ...)
}
\arguments{
\item{intervention_data}{A data frame containing individual patient data from
the intervention study.}

\item{matching_vars}{A character vector giving the names of the covariates to
use in matching. These names must match the column names in
intervention_data.}

\item{method}{The method used for optimisation - The default is method =
"BFGS". Refer to \code{\link[stats]{optim}} for options.}

\item{...}{Additional arguments to be passed to optimisation functions such
as the method for maximum likelihood optimisation. Refer to \code{\link[stats]{optim}} 
for options.}
}
\value{
A list containing 2 objects. First, a data frame named analysis_data
  containing intervention_data with additional columns named wt (weights) and
  wt_rs (rescaled weights). Second, a vector called matching_vars of the
  names of the centered matching variables used.
}
\description{
Estimate propensity weights for matching-adjusted indirect comparison (MAIC).
}
\details{
The premise of MAIC methods is to adjust for between-trial
  differences in patient demographic or disease characteristics at baseline.
  When a common treatment comparator or ‘linked network’ are unavailable, a
  MAIC assumes that differences between absolute outcomes that would be
  observed in each trial are entirely explained by imbalances in prognostic
  variables and treatment effect modifiers.

  The aim of the MAIC method is to estimate a set of propensity weights based
  on prognostic variables and treatment effect modifiers. These weights can
  be used in subsequent statistical analysis to adjust for differences in
  patient characteristics between the population in the intervention trial
  and the population in a comparator study. For additional details on the
  statistical methods, refer to the package vignette.

  The data required for an unanchored MAIC are:

  \itemize{
    \item Individual patient data from a single arm study of 'intervention'
    \item Aggregate summary data for 'comparator'. This could be from a
    single arm study of the comparator or from one arm of a randomized
    controlled trial.
    \item Psuedo patient data from the comparator study. This is not required
    for the matching process but is needed to derive the relative treatment
    effects between the intervention and comparator.
    }

    For the matching process:

    \enumerate{
      \item All binary variables to be used in the matching should be coded 1
      and 0
      \item The variable names need to be listed in a character vector called
      match_cov
      \item Aggregate baseline characteristics (number of patients, mean and
      SD for continuous variables and proportion for binary variables) from
      the comparator trial are needed as a data frame. Naming of the
      covariates in this data frame should be consistent with variable names
      in the intervention data.
      \item Patient baseline characteristics in the intervention study are
      centered on the value of the aggregate data from the comparator study
      \item The estimate_weights function can then be used to estimate
      propensity weights for each patient in the intervention study
    }

    For full details refer to the example below and the package vignette
}
\examples{

# This example code estimates weights for individual patient data from a single
# arm study of 'intervention' based on aggregate baseline characteristics from
# the comparator trial

library(dplyr)
library(MAIC)


#### Prepare the data ----------------------------------------------------------

### Intervention data

# Read in relevant ADaM data and rename variables of interest
adsl <- read.csv(system.file("extdata", "adsl.csv", package = "MAIC",
                             mustWork = TRUE))
adrs <- read.csv(system.file("extdata", "adrs.csv", package = "MAIC",
                             mustWork = TRUE))
adtte <- read.csv(system.file("extdata", "adtte.csv", package = "MAIC",
                              mustWork = TRUE))

adsl <- adsl \%>\% # Data containing the matching variables
  mutate(SEX=ifelse(SEX=="Male", 1, 0)) # Coded 1 for males and 0 for females

adrs <- adrs \%>\% # Response data
  filter(PARAM=="Response") \%>\%
  transmute(USUBJID, ARM, response=AVAL)

adtte <- adtte \%>\% # Time to event data (overall survival)
  filter(PARAMCD=="OS") \%>\%
  mutate(Event=1-CNSR) \%>\% #Set up coding as Event = 1, Censor = 0
  transmute(USUBJID, ARM, Time=AVAL, Event)

# Combine all intervention data
intervention_input <- adsl \%>\%
  full_join(adrs, by=c("USUBJID", "ARM")) \%>\%
  full_join(adtte, by=c("USUBJID", "ARM"))

# List out the variables in the intervention data that have been identified as
# prognostic factors or treatment effect modifiers and will be used in the
# matching
match_cov <- c("AGE",
               "SEX",
               "SMOKE",
               "ECOG0")

## Baseline data from the comparator trial
# Baseline aggregate data for the comparator population
target_pop <- read.csv(system.file("extdata", "aggregate_data.csv",
                                     package = "MAIC", mustWork = TRUE))

# Rename target population cols to be consistent with match_cov
target_pop_standard <- target_pop \%>\%
        rename(
           N=N,
           Treatment=ARM,
           AGE=age.mean,
           SEX=prop.male,
           SMOKE=prop.smoke,
           ECOG0=prop.ecog0
              ) \%>\%
        transmute(N, Treatment, AGE, SEX, SMOKE, ECOG0)


#### Estimate weights ----------------------------------------------------------

### Center baseline characteristics
# (subtract the aggregate comparator data from the corresponding column of
# intervention PLD)
intervention_data <- intervention_input \%>\%
    mutate(
     Age_centered = AGE - target_pop$age.mean,
     # matching on both mean and standard deviation for continuous variables (optional)
     Age_squared_centered = (AGE^2) - (target_pop$age.mean^2 + target_pop$age.sd^2),
     Sex_centered = SEX - target_pop$prop.male,
     Smoke_centered = SMOKE - target_pop$prop.smoke,
     ECOG0_centered = ECOG0 - target_pop$prop.ecog0)

## Define the matching covariates
cent_match_cov <- c("Age_centered",
                    "Age_squared_centered",
                    "Sex_centered",
                    "Smoke_centered",
                    "ECOG0_centered")

## Optimization procedure
# Following the centering of the baseline characteristics of the intervention
# study, patient weights can be estimated using estimate_weights
# The function output is a list containing (1) a data set of the individual
# patient data with the assigned weights "analysis_data" and (2) a vector
# containing the matching variables "matching_vars"
est_weights <- estimate_weights(intervention_data = intervention_data,
                                matching_vars = cent_match_cov)

}
\references{
NICE DSU TECHNICAL SUPPORT DOCUMENT 18: METHODS FOR
  POPULATION-ADJUSTED INDIRECT COMPARISONS IN SUBMSISSIONS TO NICE, REPORT BY
  THE DECISION SUPPORT UNIT, December 2016
}
\seealso{
\code{\link{optim}}
}
