% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{find_ITC}
\alias{find_ITC}
\title{Calculate relative risk using indirect treatment comparison (i.e. consistency equation) for anchored comparison}
\usage{
find_ITC(AB = NULL, AC = NULL, RR = "HR", CI.perc = 0.95)
}
\arguments{
\item{AB}{Log relative risk from AB trial (can be log hazard ratio or log odds ratio)}

\item{AC}{Log relative risk from AC trial}

\item{RR}{Specifies which type of relative risk is to be calculated. Options include "HR" for hazard ratio and "OR" for odds ratio.}

\item{CI.perc}{Defines the percentage of confidence interval needed.}
}
\value{
RR of C vs B with standard error and confidence interval
}
\description{
Convenient function to calculate relative risk using indirect treatment comparison.
Function calculates treatment effect of C vs B via subtracting treatment effect of C vs A from treatment effect of B vs A.
Function also calculates standard error and confidence interval for the relative risk (C vs B).
This function is for an anchored comparison as it assumes a common comparator "A".
}
