% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{find_RR}
\alias{find_RR}
\title{Retrieve relative risk and log relative risk with associated standard error and confidence interval}
\usage{
find_RR(fit = NULL, logRR = NULL, logRR_SE = NULL, RR = "HR", CI.perc = 0.95)
}
\arguments{
\item{fit}{Regression fit (i.e. cox regression or logistic regression) where the logRR can be extracted from}

\item{logRR}{Instead of providing the fitted model (via parameter \code{fit}), one can specify logRR manually through this parameter}

\item{logRR_SE}{If logRR is specified, standard error has to be provided.}

\item{RR}{Specifies which type of relative risk is to be calculated. Options include "HR" for hazard ratio and "OR" for odds ratio.}

\item{CI.perc}{Defines the percentage of confidence interval needed.}
}
\value{
RR and logRR with standard error and confidence interval
}
\description{
Convenient function to retrieve both the relative risk (RR) and log relative risk (logRR) either from
fitted model or separate calculation.
Relative risk would be hazard ratio for the cox regression and odds ratio for the logistic regression.
Delta method is used to calculate the standard error. 
There are two ways of finding CI: Use se(RR) obtained via the delta method and 
calculate the end points or form a confidence interval for logRR and then exponentiate 
the endpoints. We will use the second method since logRR converges to a normal distribution 
more quickly than RR. This is also a default coxph method in \code{survival} R package.
}
