% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haemoglobin}
\alias{haemoglobin}
\title{Simulated trial dataset for iron deficiency anaemia}
\format{
## `haemoglobin`
A data frame with 300 rows and 5 columns:
\describe{
  \item{arm}{Intervention arm}
  \item{sex}{"Female" or "Male"}
  \item{age}{Age at baseline (years)}
  \item{Hb_bl}{Serum haemoglobin at baseline (g/dL)}
  \item{Hb_delta}{Difference in serum haemoglobin between baseline and week 12 (g/dL)}
  }
}
\usage{
haemoglobin
}
\description{
A simulated dataset representing a comparison of investigational 
intervention "B" against control intervention "A". To be used in conjunction
with "haemoglobin_aggregate_data" for demonstrating MAIC
}
\keyword{datasets}
