% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haemoglobin_aggregate_data}
\alias{haemoglobin_aggregate_data}
\title{Aggregate data from two simulated iron deficiency anaemia trials}
\format{
## "haemoglobin_aggregate_data"
A data frame with 10 rows and 11 columns:
\describe{
  \item{N}{Number of patients}
  \item{age_mean}{mean age at baseline (years)}
  \item{age_sd}{standard deviation of age at baseline (years)}
  \item{proportion_female}{proportion of patients female}
  \item{Hb_bl_mean}{mean serum haemoglobin at baseline (g/dL)}
  \item{Hb_bl_sd}{standard deviation of serum haemoglobin at baseline (g/dL)}
  \item{Hb_delta_mean}{mean difference in serum haemoglobin between baseline and week 12 (g/dL)}
  \item{Hb_delta_sd}{standard deviation of serum haemoglobin between baseline and week 12 (g/dL)}
  \item{id}{population identifier}
  \item{difference_Hb_delta_mean}{for rows representing paired study arms, difference in mean difference in serum haemoglobin between baseline and week 12 (g/dL)}
  \item{difference_Hb_delta_sd}{for rows representing paired study arms, standard deviation of difference in mean difference in serum haemoglobin between baseline and week 12 (g/dL)}
  }
}
\usage{
haemoglobin_aggregate_data
}
\description{
Summary baseline characteristics and outcomes from two simulated trials - 
"AB" measuring intervention "B" versus "A" per the data in "haemoglobin"
and "AC" measuring intervention "C" versus "A". Counterfactual outcomes are
provided for both trials - in "AB", replacing "B" with "C", and in "AC" 
replacing "C" with "B"
}
\keyword{datasets}
