% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_survival.R
\name{HR_summary}
\alias{HR_summary}
\title{Summary table of hazard ratio estimates}
\usage{
HR_summary(
  analysis_data,
  bootstrap = TRUE,
  bootstrap_method = "percentile",
  ...
)
}
\arguments{
\item{analysis_data}{A data frame containing data with weights (derived from
\code{\link{analysis_dataset}}).}

\item{bootstrap}{Logical. If TRUE (Default) perform bootstrapping of the
propensity weights.}

\item{...}{Additional arguments to be passed to the bootstrapping functions.
For example the number of simulations.}

\item{bootstrap_method.}{A character string specifying the bootstrap method
to be used. Default = 'percentile'. This is ignored if bootstrap = FALSE.}
}
\value{
A summary table of hazard ratio estimates with 95\% CI
}
\description{
Returns table of hazard ratios with 95\% CI: no weighting (naive Cox
proportional hazards model), weighted Cox model using fixed weights and a
weighted Cox model using bootstrapped weights
}
\seealso{
\code{\link{analysis_dataset}}, \code{\link{bootstrap_HR}}
}
