% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_survival.R
\name{KM_summary}
\alias{KM_summary}
\title{Summarize Kaplan-Meier estimate of survival}
\usage{
KM_summary(survfit_object)
}
\arguments{
\item{survfit_object}{A survival model object derived from
\code{\link{survfit}}. This can be either a weighted or unweighted estimate.}
}
\value{
A summary table of the Kaplan-Meier estimate including number of patients, number of
events and median time to event
}
\description{
Return a summary of the Kaplan-Meier estimate of survival based on
\code{\link[survival]{survfit}} including number of patients, number of
events and median time to event
}
