% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIC.R
\docType{package}
\name{MAIC}
\alias{MAIC}
\title{MAIC: A package for performing matched-adjusted indirect comparisons}
\description{
The MAIC package provides functions to help perform and summarize results
from matched-adjusted indirect comparisons
}
