% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maic.R
\name{MaicInput}
\alias{MaicInput}
\title{Constructor for a maic.input object}
\usage{
MaicInput(n.adjustments, n.matches, excluded, input.matrix)
}
\arguments{
\item{n.adjustments}{Numeric, number of variables that have had any
comparison performed}

\item{n.matches}{Numeric, number of matching variables}

\item{excluded}{Logical vector; which index rows have been excluded from 
matching}

\item{input.matrix}{Numeric matrix, centred MAIC input matrix}
}
\value{
An object of class \code{maic.input}
}
\description{
Constructor for a maic.input object
}
