% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_binary.R
\name{OR_summary}
\alias{OR_summary}
\title{Summary table of odds ratio estimates}
\usage{
OR_summary(
  analysis_data,
  bootstrap = TRUE,
  bootstrap_method = "percentile",
  ...
)
}
\arguments{
\item{analysis_data}{A data frame containing data with weights (derived from
\code{\link{analysis_dataset}}).}

\item{bootstrap}{Logical. If TRUE (Default) perform bootstrapping of the
propensity weights.}

\item{bootstrap_method}{A character string specifying the bootstrap method
to be used. Default = 'percentile'. This is ignored if bootstrap = FALSE.}

\item{...}{Additional arguments to be passed to the bootstrapping functions.
For example the number of simulations.}
}
\value{
A data frame of ORs with 95\% CI.
}
\description{
Returns table of odds ratios with 95\% CI: no weighting (naive logistic
regression), weighted logistic regression model using fixed weights and a
weighted logistic regression using bootstrapped weights
}
\seealso{
\code{\link{analysis_dataset}}, \code{\link{bootstrap_OR}}
}
