% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data}
\usage{
preprocess_data(intervention_input, target_pop)
}
\arguments{
\item{intervention_input}{A data frame containing individual patient data from
the intervention study.}

\item{target_pop}{A data frame containing aggregate dataset for the target population. 
Variables are followed by one of the following suffixes to denote the type of summary: 
varname_mean, varname_sd, varname_median, varname_prop. 
After preprocessing these summary suffixes, intervention_input is 
centered using the aggregate data averages.}
}
\value{
A list containing 2 objects. First, a data frame named analysis_data
  containing intervention_data with additional columns named wt (weights) and
  wt_rs (rescaled weights). Second, a vector called matching_vars of the
  names of the centered matching variables used.
}
\description{
Preprocess data before estimating weights for matching-adjusted indirect comparison (MAIC).
}
\references{
NICE DSU TECHNICAL SUPPORT DOCUMENT 18: METHODS FOR
  POPULATION-ADJUSTED INDIRECT COMPARISONS IN SUBMSISSIONS TO NICE, REPORT BY
  THE DECISION SUPPORT UNIT, December 2016
}
