% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_baselines.R
\name{summarize_baseline_variance}
\alias{summarize_baseline_variance}
\title{Summarize ratio of variance of baseline characteristics before and after weighting}
\usage{
summarize_baseline_variance(analysis_data, matching_vars)
}
\arguments{
\item{analysis_data}{A data frame containing data with weights (derived from
\code{\link{analysis_dataset}}).}

\item{matching_vars}{A character vector giving the names of the covariates used
in matching.}
}
\value{
Table of the ratio of variance before and after matching
}
\description{
Summarize ratio of variance of baseline characteristics before and after
weighting. The purpose of this metric is to assess whether re-weighting has
affected the shape of the distribution as well as the location of the
distribution for each matching variable.
}
\seealso{
\code{\link{analysis_dataset}}
}
