% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_baselines.R
\name{summarize_baselines}
\alias{summarize_baselines}
\title{Summarize baseline characteristics before and after weighting}
\usage{
summarize_baselines(analysis_data, matching_vars)
}
\arguments{
\item{analysis_data}{A data frame containing data with propensity weights
(derived from \code{\link{analysis_dataset}}).}

\item{matching_vars}{A character vector giving the names of the covariates used
in matching.}
}
\value{
A table with three rows summarising the baseline characteristics for
  weighted intervention data, unweighted intervention data and comparator
  data.
}
\description{
Summarize baseline characteristics before and after weighting
}
\seealso{
\code{\link{analysis_dataset}}
}
