% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_pop_standard}
\alias{target_pop_standard}
\title{target_pop_standard}
\format{
A data frame with 1 row and 6 variables:
\describe{
  \item{N}{Total number of patients in the comparator population}
  \item{Treatment}{Treatment received in the comparator population}
  \item{AGE}{Mean age}
  \item{SEX}{Proportion of males}
  \item{SMOKE}{Proportion of patients who smoke}
  \item{ECOG0}{Proportion of patients with ECOG 0}


}
}
\usage{
target_pop_standard
}
\description{
A data frame containing the aggregate baseline characteristics in the
comparator population.
}
\keyword{datasets}
