% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_survival.R
\name{bootstrap_HR}
\alias{bootstrap_HR}
\title{Bootstrapping for MAIC propensity weighted hazard ratios}
\usage{
bootstrap_HR(analysis_dataset, n_sim = 1000)
}
\arguments{
\item{n_sim}{Number of simulations. Default = 1000.}

\item{analysis_data}{A data frame containing data with weights (derived from
\code{\link{analysis_dataset}}).}
}
\value{
A data frame of n_sim bootstrapped HR values.
}
\description{
Bootstrapping of the hazard ratio to capture the uncertainty in the
estimation of MAIC propensity weights using the following general process:
\enumerate{
   \item Sample patients with replacement from the intervention study
   \item Estimate propensity weights for the current sample of patients
   \item Calculate a weighted hazard ratio
   \item Repeat the steps above N times. Approx. 1000 simulations are usually
   sufficient
}
Refer to \code{\link{HR_summary}} for alternative approaches to utilising
bootstrapped samples
}
\seealso{
\code{\link{analysis_dataset}}, \code{\link{estimate_weights}}, \code{\link{HR_summary}}
}
