% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_HR}
\alias{bootstrap_HR}
\title{Bootstrapping for MAIC propensity weighted hazard ratios HELP PAGE NEEDS UPDATING}
\usage{
bootstrap_HR(intervention_data, comparator_data, matching, i, model)
}
\arguments{
\item{intervention_data}{A data frame containing data with weights (derived from
\code{\link{est_weights}}).}

\item{model}{A model formula in the form 'surv(Time ~ Event)'.
Variable names need to match the corresponding columns in intervention_data}
}
\value{
A HR value.
}
\description{
Bootstrapping of the hazard ratio to capture the uncertainty in the
estimation of MAIC propensity weights using the following general process:
bootstrapped samples
}
\seealso{
\code{\link{analysis_dataset}}, \code{\link{estimate_weights}},
}
