% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_OR}
\alias{bootstrap_OR}
\title{Bootstrapping for MAIC weighted odds ratios}
\usage{
bootstrap_OR(intervention_data, comparator_data, matching, i, model)
}
\arguments{
\item{intervention_data}{A data frame containing containing individual patient data from the intervention study.}

\item{comparator_data}{A data frame containing pseudo individual patient data from the comparator study.
The outcome variables names must match intervention_data.}

\item{matching}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in intervention_data
and comparator_data.}

\item{i}{Index used to select a sample within \code{\link{boot}}.}

\item{model}{A model formula in the form 'endpoint ~ treatment_var'.
Variable names need to match the corresponding columns in intervention_data.}
}
\value{
The OR as a numeric value.
}
\description{
A function required for the "statistic" argument in the \code{\link{boot}} function.
Performs MAIC weighting using {\link{estimate_weights}} and returns a weighted odds ratio (OR) from a binomial generalised linear model.
}
\seealso{
\code{\link{estimate_weights}}, \code{\link{boot}}
}
