% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_binary.R
\name{bootstrap_OR}
\alias{bootstrap_OR}
\title{Bootstrapping for MAIC propensity weighted odds ratios}
\usage{
bootstrap_OR(analysis_data, n_sim = 1000)
}
\arguments{
\item{analysis_data}{A data frame containing data with weights (derived from
\code{\link{analysis_dataset}}).}

\item{n_sim}{Number of simulations. Default = 1000.}
}
\value{
A data frame n_sim bootstrapped OR values
}
\description{
Bootstrapping of the odds ratio to capture the uncertainty in the
estimation of MAIC propensity weights using the following general process:
\enumerate{
   \item Sample patients with replacement from the intervention study
   \item Estimate propensity weights for the current sample of patients
   \item Calculate a weighted odds ratio
   \item Repeat the steps above N times. Approx. 1000 simulations are usually
   sufficient
}
Refer to \code{\link{OR_summary}} for alternative approaches to utilising
bootstrapped samples
}
\seealso{
\code{\link{analysis_dataset}}, \code{\link{estimate_weights}}, \code{\link{OR_summary}}
}
