% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{check_weights}
\alias{check_weights}
\title{Checking whether optimization has worked}
\usage{
check_weights(
  analysis_data = NULL,
  matching_vars = NULL,
  target_pop_standard = NULL
)
}
\arguments{
\item{analysis_data}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using estimate_weights).}

\item{target_pop_standard}{aggregate characteristics of the comparator trial with the same naming as the analysis_data}

\item{match_covs}{A character vector giving the names of the covariates that were used to estimate weights}
}
\value{
Summary of patient characteristics before and after matching, including ESS and comparator trial aggregate summary
}
\description{
Convenient function to check whether the re-weighted baseline characteristics for
the intervention-treated patients match those aggregate characteristics from the
comparator trial and outputs a summary that can be used for reporting
}
\examples{

# This example code uses the weighted individual patient data, outputted from
# the estimate_weights function to perform weight diagnostics. The weighted data
# is saved within est_weights. To check the weighted aggregate baseline
# characteristics for 'intervention' match those in the comparator data,
# standardized data "target_pop_standard" is used. Please see the package
# vignette for more information on how to use the estimate_weights function and
# derive the "target_pop_standard" data.

library(dplyr)
library(MAIC)

# load est_weights
data(est_weights, package = "MAIC")

# load target_pop_standard
data(target_pop_standard, package = "MAIC")

# List out the uncentered variables used in the matching
match_cov <- c("AGE",
               "SEX",
               "SMOKE",
               "ECOG0")

# Are the weights sensible? ----------------------------------------------------

# The wt_diagnostics function requires the output from the estimate_weights
# function and will output:
# - the effective sample size (ESS)
# - a summary of the weights and rescaled weights (mean, standard deviation,
#   median, minimum and maximum)
# - a unique set of weights with the corresponding patient profile based on the
#   matching variables

diagnostics <- wt_diagnostics(est_weights$analysis_data,
                              vars = match_cov)
diagnostics$ESS
diagnostics$Summary_of_weights
diagnostics$Weight_profiles

# Each of the wt_diagnostics outputs can also be estimated individually
ESS <- estimate_ess(est_weights$analysis_data)
weight_summ <- summarize_wts(est_weights$analysis_data)
wts_profile <- profile_wts(est_weights$analysis_data, vars = match_cov)

# Plot histograms of unscaled and rescaled weights
# bin_width needs to be adapted depending on the sample size in the data set
histogram <- hist_wts(est_weights$analysis_data, bin = 50)
histogram


# Has the optimization worked? -------------------------------------------------

# The following code produces a summary table of the intervention baseline
# characteristics before and after matching compared with the comparator
# baseline characteristics:

check_weights(analysis_data = est_weights$analysis_data, matching_vars = match_cov, 
                          target_pop_standard = target_pop_standard)
}
\seealso{
\code{\link{estimate_weights}}
}
