% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{estimate_ess}
\alias{estimate_ess}
\title{Estimate effective sample size HELP PAGE NEEDS UPDATING}
\usage{
estimate_ess(data, wt = "wt")
}
\arguments{
\item{intervention_wts}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using estimate_weights).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is WT.}
}
\value{
The effective sample size (ESS) as a numeric value.
}
\description{
Estimate the effective sample size (ESS).
}
\references{
NICE DSU TECHNICAL SUPPORT DOCUMENT 18: METHODS FOR
  POPULATION-ADJUSTED INDIRECT COMPARISONS IN SUBMSISSIONS TO NICE, REPORT BY
  THE DECISION SUPPORT UNIT, December 2016
}
\seealso{
\code{\link{estimate_weights}}
}
