% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate MAIC propensity weights}
\usage{
estimate_weights(intervention_data, matching_vars)
}
\arguments{
\item{intervention_data}{A data frame containing individual patient data from the intervention study.}

\item{matching_vars}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in intervention_data.}

\item{...}{Additional arguments to be passed to optimisation functions such
as the method for maximum likelihood optimisation. The default is method =
"BFGS". Refer to \code{\link[stats]{optim}} for options.}
}
\value{
A list containing 2 objects. First, a data frame named analysis_data containing intervention_data
  with additional columns named wt (weights) and wt_rs (rescaled weights).
  Second, a vector called matching_vars of the matching variables names used.
}
\description{
Estimate propensity weights for matching-adjusted indirect comparison (MAIC).
}
\references{
NICE DSU TECHNICAL SUPPORT DOCUMENT 18: METHODS FOR
  POPULATION-ADJUSTED INDIRECT COMPARISONS IN SUBMSISSIONS TO NICE, REPORT BY
  THE DECISION SUPPORT UNIT, December 2016
}
\seealso{
\code{\link{optim}}
}
