% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate MAIC propensity weights}
\usage{
estimate_weights(intervention_data, matching_vars, ...)
}
\arguments{
\item{intervention_data}{A data frame containing individual patient data from the intervention study.}

\item{matching_vars}{A character vector giving the names of the covariates to use
in matching. These names must match the column names in intervention_data
and baseline_comparator.}

\item{...}{Additional arguments to be passed to optimisation functions such
as the method for maximum likelihood optimisation. The default is method =
"BFGS". Refer to \code{\link[stats]{optim}} for options.}

\item{baseline_comparator}{A data frame of average covariate values from the comparator study.}
}
\value{
A data frame named intervention_wts including the intervention data
  with additional columns named WT (weights) and WT_RS (rescaled weights).
}
\description{
Estimate propensity weights for matched-adjusted indirect comparison (MAIC)
}
\references{
NICE DSU TECHNICAL SUPPORT DOCUMENT 18: METHODS FOR
  POPULATION-ADJUSTED INDIRECT COMPARISONS IN SUBMSISSIONS TO NICE, REPORT BY
  THE DECISION SUPPORT UNIT, December 2016
}
