% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate MAIC propensity weights}
\usage{
estimate_weights(
  intervention_data,
  match_cov,
  startVal = 0,
  method = "BFGS",
  ...
)
}
\arguments{
\item{intervention_data}{A data frame containing individual patient data from 
the intervention study. Intervention_data is assumed to have been preprocessed using 
preprocess_data (i.e. centered using aggregate data means)}

\item{match_cov}{A character vector giving the names of the covariates to
use in matching. These names must match the column names in intervention_data.}

\item{startVal}{a scalar, the starting value for all coefficients of the propensity score 
regression}

\item{method}{The method used for optimisation - The default is method =
"BFGS". Refer to \code{\link[stats]{optim}} for options.}

\item{...}{Additional arguments to be passed to optimisation functions such
as the method for maximum likelihood optimisation. Refer to \code{\link[stats]{optim}} 
for options.}
}
\value{
A list with 4 elements
 \item{wt} a numeric vector of unscaled individual weights.
 \item{wt.rs} a numerical vector of rescaled individual weights, with summation equaling to sample size (# rows of input \code{EM})
 \item{ess} effective sample size, square of sum divided by sum of squares
 \item{opt} R object returned by \code{base::optim()}, for assess convergence and other details
}
\description{
Estimate propensity weights for matching-adjusted indirect comparison (MAIC).
}
\references{
NICE DSU TECHNICAL SUPPORT DOCUMENT 18: METHODS FOR
POPULATION-ADJUSTED INDIRECT COMPARISONS IN SUBMSISSIONS TO NICE, REPORT BY
THE DECISION SUPPORT UNIT, December 2016
}
\seealso{
\code{\link{optim}}
}
