% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{find_RR}
\alias{find_RR}
\title{Calculate relative risk with standard error and confidence interval from log relative risk (i.e. log hazard ratio or log odds ratio) of the fitted model}
\usage{
find_RR(fit = NULL, logRR = NULL, logRR_SE = NULL, RR = "HR", CI.perc = 0.95)
}
\arguments{
\item{fit}{Regression fit (i.e. cox regression or logistic regression) where the logRR can be extracted from}

\item{logRR}{Instead of providing the fitted model (via parameter \code{fit}), one can specify logRR manually through this parameter}

\item{logRR_SE}{If logRR is specified, standard error has to be provided.}

\item{RR}{Specifies which type of relative risk is to be calculated. Options include "HR" for hazard ratio and "OR" for odds ratio.}

\item{CI.perc}{Defines the percentage of confidence interval needed.}
}
\value{
RR with standard error and confidence interval
}
\description{
Convenient function to calculate relative risk (RR) from log relative risk (logRR).
Relative risk would be hazard ratio for the cox regression and odds ratio for the logistic regression.
This function uses delta method to calculate the standard error and confidence interval for the RR.
}
