% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{hist_wts}
\alias{hist_wts}
\title{Produce histograms of weights and rescaled weights}
\usage{
hist_wts(data, wt_col = "wt", rs_wt_col = "wt_rs", bin = 30)
}
\arguments{
\item{data}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using \code{\link{estimate_weights}}).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is wt.}

\item{rs_wt_col}{The name of the rescaled weights column in the data frame
containing the intervention individual patient data and the MAIC propensity
weights. The default is wt_rs.}

\item{bin}{Number of bins to plot histogram. The default is 30.}
}
\value{
A histogram plot of the weights and rescaled weights.
}
\description{
Produce a plot containing two histograms (one of the weights and one of the rescaled weights).
}
\seealso{
\code{\link{estimate_weights}}
}
