% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{profile_wts}
\alias{profile_wts}
\title{Produce a data frame of the weights assigned to alternative patient profiles HELP PAGE NEEDS UPDATING}
\usage{
profile_wts(data, wt_col = "wt", wt_rs = "wt_rs", vars)
}
\arguments{
\item{data}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using \code{\link{estimate_weights}}).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is wt.}

\item{vars}{A character vector giving the variable names of the baseline characteristics. These names must match the column names in
data.}

\item{rs_wt_col}{The name of the rescaled weights column in the data frame
containing the intervention individual patient data and the MAIC propensity
weights. The default is wt_rs}
}
\value{
A data frame that includes a summary of patient characteristics
  associated with alternative weight values ordered by ascending wt_col.
}
\description{
Select the patient characteristics used in the matching and the MAIC weights
and output a data frame of unique propensity weight values with the
associated summary baseline characteristics. This data frame helps to
understand how different patient profiles are contributing to the analyses by
illustrating the patient characteristics associated with different weight
values. For example, min, max and median weights. This function is most
useful when only matching on binary variables as there are fewer unique
values.
}
\seealso{
\code{\link{estimate_weights}}
}
