% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{summarize_wts}
\alias{summarize_wts}
\title{Summarize the weight values}
\usage{
summarize_wts(data, wt_col = "wt", rs_wt_col = "wt_rs")
}
\arguments{
\item{data}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using \code{\link{estimate_weights}}).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is wt.}

\item{rs_wt_col}{The name of the rescaled weights column in the data frame
containing the intervention individual patient data and the MAIC propensity
weights. The default is wt_rs.}
}
\value{
A data frame that includes a summary (minimum, maximum, median, mean) of the weights and rescaled weights.
}
\description{
Produce a summary of the weights (minimum, maximum, median, mean, SD). Mean
and standard deviation are provided for completeness. In practice the
distribution of weights may be skewed in which case mean and SD should be
interpreted with caution.
}
\seealso{
\code{\link{estimate_weights}}
}
