% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{summarize_wts}
\alias{summarize_wts}
\title{Summarize the weight values HELP PAGE NEEDS UPDATING}
\usage{
summarize_wts(data, wt = "wt")
}
\arguments{
\item{intervention_wts}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using estimate_weights).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is WT.}
}
\value{
A data frame that inlcudes a summary (minimum, maximum, median, mean) of the weights.
}
\description{
Produce a summary of the weights (minimum, maximum, median, mean, sd). Mean
and standard deviation are provided for completeness. In practice the
distribution of weights may be skewed in which case mean and sd should be
interpreted with caution
}
\seealso{
\code{\link{estimate_weights}}
}
