% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{wt_diagnostics}
\alias{wt_diagnostics}
\title{Weight diagnostics}
\usage{
wt_diagnostics(data, wt_col = "wt", wt_rs = "wt_rs", vars)
}
\arguments{
\item{data}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using estimate_weights).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is 'wt'.}

\item{vars}{A character vector giving the variable names of the baseline characteristics. These names must match the column names in
data.}

\item{rs_wt_col}{The name of the rescaled weights column in the data frame
containing the intervention individual patient data and the MAIC propensity
weights. The default is wt_rs}
}
\value{
List of the following:
\itemize{
  \item The effective sample size (ESS) as a numeric value.
  \item A data frame that includes a summary (minimum, maximum, median, mean) of the weights.
  \item A data frame that includes a summary of patient characteristics
  associated with each weight value.
}
}
\description{
Produce a set of useful diagnostic metrics to summarize propensity weights
\itemize{
  \item ESS (\code{\link{estimate_ess}})
  \item Summary statistics of the weights: minimum, maximum, median, mean, SD (\code{\link{summarize_wts}})
  \item Patient profile associated with weight values (\code{\link{profile_wts}})
}
}
\seealso{
\code{\link{estimate_weights}}, \code{\link{estimate_ess}}, \code{\link{summarize_wts}}, \code{\link{profile_wts}}
}
