% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{wt_diagnostics}
\alias{wt_diagnostics}
\title{Weight diagnostics}
\usage{
wt_diagnostics(intervention_wts, wt_col = WT, rs_wt_col = WT_RS, matching_vars)
}
\arguments{
\item{intervention_wts}{A data frame containing individual patient data from
the intervention study, including a column containing the weights (derived
using estimate_weights).}

\item{wt_col}{The name of the weights column in the data frame containing the
intervention individual patient data and the MAIC propensity weights. The
default is WT.}

\item{rs_wt_col}{The name of the rescaled weights column in the data frame
containing the intervention individual patient data and the MAIC propensity
weights. The default is WT_RS.}

\item{matching_vars}{A character vector giving the names of the covariates to
use in matching.}
}
\value{
A list of data frames to use for weight diagnostics.
}
\description{
Produce a set of useful diagnostic metrics to summarize propensity weights
\itemize{
  \item ESS
  \item Summary statistics of the weights: minimum, maximum, median, mean, sd
  \item Patient profile associated with given weight values
}
}
\details{
This function calls each of estimate_ess, summarize_wts and profile_wts in turn.
}
\seealso{
\code{\link{estimate_weights}}, \code{\link{estimate_ess}}, \code{\link{summarize_wts}}, \code{\link{profile_wts}}
}
